disable the 140 chars counting before posting a tweet
there is a bug where non-English messages are limited to less than 140 characters because unicode characters are counted as more than 1 character.
the simplistic, yet working, solution is to remove the client side restriction for a message to have 140 and send it anyway.
that way longer than 140 characters messages will be rejected by twitter server rather than in the client (which makes sense, the client shouldn't be enforcing server rules at the first place TMHO)