fixes issue #294.
Errors were from the way that http requests were handled with urllib2.urlopen, which returns a not-really-file like object (as opposed to the python3 version, which does). Additionally, there were some issues with how media_ids was joined into a string for posting in PostUpdate; those have been fixed. Tests have been added for most of these features with the exception of PostUpdate with http because of the way responses/mock works with a streaming URL.