removes check on reset time.
The idea is that if you are making a call to, say, ``/statuses/lookup.json`` and the limit information is stale, i.e., limit.reset is in the past, then you should make another call to ``/application/rate_limit_status.json`` to get your new limits and reset.

That said, this could result in a lot of hits to *that* endpoint, thereby rate limiting your application on the ``/application/rate_limit_status.json`` endpoint and causing unnecessary sleeps for the API. Worst case scenario is that your application calls a bunch of different methods hitting a bunch of different endpoints and you end up making a ton of checks on rate_limit_status.json and, since ``Api.InitializeRateLimit()`` does not abide by rate limits, you could get blacklisted.