Prioritize kwargs over data when creating Status
Because the construction of a Status model prioritized
data in the `data` arg passed to NewFromJsonDict(),
the kwargs are being overwritten, which seems a little
backwards since if you specifically pass a kwarg,
I'd think that should get priority.  This doesn't change
any of behavior of the tests **except the following**:

If a tweet had a `media` node in `data['entities']`,
that was getting overwritten by the
`data['extended_entities']['media']` node. This behavior
was not documented anywhere and seems like a bug in a
few cases (i.e., when there's a photo preview and
also a video - the photo gets overwritten).