add support for Collections
Purpose
-------
Adds three methods for interacting with Collections on Twitter:

1. Api.GetCollectionEntries
   (https://dev.twitter.com/rest/reference/get/collections/entries) will
return a Collection populated with Status objects found in that
Collection. This Collection object will have a not-None `statuses`
attribute. This Collection will also have a `timeline` attribute for
organizing tweets.

2. Api.GetCollection
   (https://dev.twitter.com/rest/reference/get/collections/show) will
return a Collection with some information about that collection such as
name, id, user, etc.

3. Api.GetCollectionList
   (https://dev.twitter.com/rest/reference/get/collections/list) will
return a list of Collections (along with previous and next cursors)
either for a given: user_id, screen_name, or containing a particular
status.

Additionally, a Collection model has been added. Because the endpoints
above return a couple different flavors of Collections (with and without
statuses/timelines/etc.), users should be aware that attributes may or
may not exist. As with all models, there is a `_json` attribute if it is
required.

Testing
-------
Full coverage on all new methods. Test data has been added in
testdata/collections. Tests are located in tests/test_collections.py