move UserStatus.connections to private class attr
bc US.connections is a public variable, there
is some confusion regarding its use. the original
idea was that US.connections would only hold the
default connection types available, much like
the param_defaults attribute for other models;
however, bc of the name, it was implied that this
attribute contained the connections for the specific
instance, resulting in confusion between it and
the intended attributes (i.e., us.followed_by etc).

this commit moves the attribute to a private
attribute (._connections) so that it is clearer
to end-users that it should not be used for a
specific instance.

additionally, this creates a @property with the
class-specific attributes for US.following, etc.