Fix typing issue in `Api._ParseAndCheckTwitter`
Right now whenever the twitter API returns an un-parsable response that isn't explicitly known/handled, the TwitterException sets the message parameter to a `set` type, not a `dict` type or a `str` type (which is used in some cases elsewhere). 

This is causing errors for me as I'm inspecting this exception for graceful error handling, and received a `set` much to my surprise! Looks to me like a copy/paste bug. 

I'd like to have it be a `dict`, but if you think a `str` is more appropriate that works too! 

Thanks for the great library :)